Conditionally save accessed addresses
Updating the accessed_addresses set is taking significant time from
benchmarks, between 15% and 25%.
This is only used for proof mode which, while an important use case, is
not always on and currently not supported.
The patch conditions whether we produce the addresses and update the set
to existence of the set in the VirtualMachine instance, using itself as
flag.
Because the feature is not yet implemented, the constructor just sets it
to None for simplicity, with tests explicitly overriding this value.