Add integration tests for math.cairo (#320)
* Add --cairo-path to depend on other programs

* Add integration tests and benchmark for math.cairo

The test is a function parameterized by the number of elements to test
with. There's one or more functions testing each of the functionalities
of the library on different arrays.
Division and splits are still missing and the arrays are still rather
simplistic.
The benchmark calls this function with a bigger array size to stress the
VM a little bit more.