Convert `BuiltinRunner` from trait into enum
Implementing BuiltinRunner as trait meant storing `&dyn trait` fields in
`VirtualMachine`, which makes it `!Send`, in turn making it impossible
to release the GIL in `cairo-rs-py` without wrapping it into
`Arc<Mutex<_>>` in `cairo-rs`, losing some of the compile-time benefits
of Rust and making it inconvenient to use for downstream projects.
Under the assumption that no runners will be added outside of the scope
of `cairo-rs` it was, thus, decided to just use an enum that forwards
the calls to its variants.