Fix EC op (#754)
* Add test for misdetection of points not in curve

* Fix EC op detection of points outside the curve

After the conversion to `Felt`, the points in those cells had to be
converted to `BigInt` for the functions in math_utils. But `.to_bigint`
actually converts it to the signed felt form, rather than just
extracting the value as `BigInt`.
After this patch it's hacked around by doing
`.to_biguint().to_bigint()`.