Recalibrate benchmarks for quick profiling (#923)
Adjust some parameters in cairo benchmark programs to get their runtime
in the range of 1-2s to help more quickly test optimizations and profile
execution.
These times are not optimal for comparing with CPython and PyPy, as it
doesn't give PyPy enough warmup time. A later PR may address this by
improving parameterization and separating tests for it with largar
parameters.