Reduce clones in bottleneck
`get_perm_range_check_limits` clones for immediate operands consume
about 6% of runtime in proof mode. This clone is superfluous, as the
value itself is ignored.
After this patch the position in memory is simply validated to be a felt
and then a `Some(Felt252::zero())` is passed to `decode_instruction`.