test: unit tests for cairo-vm-cli
This splits functionality into a `run` function that receives the
iterator of arguments, and `main` that simply calls it with
`std::env::args`.
Then, adds several test cases with `rstest`. It checks all combinations
of successful runs and a few of invalid arguments and missing files.
This should make the codecov path pipeline happier :)

NOTE: the dynamic layout is too slow, and it caused a panic
(subsequently fixed) as well. It is commented out from the tests until
optimized.