Add pipeline to enforce updating CHANGELOG.md
As we've been failing to keep this critical documentation up-to-date,
this acts as a strong reminder that it should be updated by failing the
checks on every PR lacking this update.
It includes the exception for the labels "pipelines" and "coverage", as
PRs that only work on those don't introduce user-visible changes.