ci: add Hyperfine pipelines (#965)
* ci: disable IAI for PRs

We're migrating to a Hyperfine based pipeline to speed up development.

* ci: add Hyperfine pipelines

Add new pipeline that runs both the base and HEAD commit one against
each other under Hyperfine with all programs compiled from the base
commit, and any changed or added programs from HEAD only with the binary
compiled from HEAD.
It creates or updates a comment with the results.