perf: specialize `u64` with `Felt252` addition (#938)
* perf: specialize add for usize and Felt252

This avoids the creation of a `Felt252` instance in `update_pc` which
takes almost 25% of the runtime in some benchmarks in proof mode.
It takes into account negative field elements, as those are the main
reason we needed to upgrade the numbers rather than extract a `usize`
value from the offset.

* Fix tests

---------

Co-authored-by: Pedro Fontana <fontana.pedro93@gmail.com>