fix: substraction between relocatables should be signed (#1218)
* fix: substraction between relocatables should be signed

It is valid as per the reference implementation to substract two
relocatables `a - b` where `b.offset > a.offset`.
The fix consists in not using the trait as implemented for `Relocatable`
(where it represents a positive distance between pointers) but to
promote the offsets to `i128` and only substract them for
`MaybeRelocatable`, as those are the language-level types.

* chore: remove uneeded stuff

* Add fix holes to if_reloc_equal.cairo test

* chore: changelog

---------

Co-authored-by: Pedro Fontana <fontana.pedro93@gmail.com>