chore: bump starknet-crypto to v0.6.1
The new version contains a bug fix on ECDSA signature verification.
Previously, the `s` range was incorrectly enforced to be lower than
EC order instead of element bound. While this did not allow invalid
signatures to be used, it was theoretically possible (1 in 2^48
signatures) to have a signature that passes `verify` yet cannot be
proven (due to being out of element bound).