Feature: compute program hash chain (#1647)
* Feature: compute program hash chain

Problem: computing the hash of a program is useful to verify its
integrity. This hash is also used for different purposes, like when
loading a program with the bootloader.

Solution: add a `program_hash` module and the
`compute_program_hash_chain` function that computes the hash of a
stripped program, making it usable with `Program` and `CairoPie`
objects.

* compile module in no-std