feat: specify initial value for exec_scopes
Context: bootloader and Starknet OS support.

Added an `exec_scopes` parameter to `cairo_run_program` to pass
variables to the hints of the program being run.

We need a way to set initial variables inside the execution scopes to
pass bootloader/OS arguments. The Python VM relies on a program input file
and uses hints to deserialize this input. In most cases we do not want
to have to deserialize these input types, but rather pass them directly
as variables inside the root execution scope.

We did not extend this feature to `cairo_run` to keep the PR minimal for
discussion, TBD if it is desirable.